refactored package to container.tools.kubectl
