+++ kubernetes/core/src/main/kotlin/com/google/kubernetes/tools/kubectl/KubectlUiResources.kt
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package com.google.kubernetes.tools.core
+package com.google.kubernetes.tools.kubectl
 
 import com.google.common.annotations.VisibleForTesting
 import com.intellij.CommonBundle
@@ -30,8 +30,8 @@ private const val BUNDLE_NAME = "messages.KubectlBundle"
  */
 @VisibleForTesting
 fun message(
-    @PropertyKey(resourceBundle = BUNDLE_NAME) key: String,
-    vararg params: String
+        @PropertyKey(resourceBundle = BUNDLE_NAME) key: String,
+        vararg params: String
 ): String {
     return CommonBundle.message(ResourceBundle.getBundle(BUNDLE_NAME), key, params)
 }
